+++ stem/control.py
@@ -1265,8 +1265,12 @@ class Controller(BaseController):
     """
     get_exit_policy(default = UNDEFINED)
 
-    Effective ExitPolicy for our relay. This accounts for
-    ExitPolicyRejectPrivate and default policies.
+    Effective ExitPolicy for our relay.
+
+    .. versionchanged:: 1.7.0
+       Policies retrieved through 'GETINFO exit-policy/full' rather than
+       parsing the user's torrc entries. This should be more reliable for
+       some edge cases. (:trac:`25739`)
 
     :param object default: response if the query fails
 
@@ -1281,23 +1285,13 @@ class Controller(BaseController):
     """
 
     with self._msg_lock:
-      config_policy = self._get_cache('exit_policy')
+      policy = self._get_cache('exit_policy')
 
-      if not config_policy:
-        policy = []
+      if not policy:
+        policy = stem.exit_policy.ExitPolicy(*self.get_info('exit-policy/full').splitlines())
+        self._set_cache({'exit_policy': policy})
 
-        if self.get_conf('ExitPolicyRejectPrivate') == '1':
-          policy.append('reject private:*')
-
-        for policy_line in self.get_conf('ExitPolicy', multiple = True):
-          policy += policy_line.split(',')
-
-        policy += self.get_info('exit-policy/default').split(',')
-
-        config_policy = stem.exit_policy.get_config_policy(policy, self.get_info('address', None))
-        self._set_cache({'exit_policy': config_policy})
-
-      return config_policy
+      return policy
 
   @with_default()
   def get_ports(self, listener_type, default = UNDEFINED):